Use semicolon instead of newline to minify
Minification aims to shorten code and reduce the number of bytes transferred.
Both "\n" and semicolons are 1 byte each, but in environments like Windows,
where "\r\n" is used for newline, a newline is 2 bytes.
Therefore, newlines are replaced with semicolons to reduce the byte count to 1 byte.

While newlines can improve the readability of backtrace in case of error,
and reducing byte count with one-liner is a trade-off,
the primary goal of minification is to reduce byte count.
Hence, reducing the byte count is prioritized.

Additionally, to comply with POSIX specification, the newline before EOF is preserved.