Prevent a syntax error when using comparison operator in the condition of ternary oprator
This commit prevents a syntax error when using comparison operator in the condition of ternary oprator

Original:

```ruby
foo == bar ? x : y
```

Before:

```ruby
foo==bar?x: y # syntax error
```

After:

```ruby
foo==bar ? x: y # OK
```

NOTE: In a Lexer, it is necessary to backtrack an indefinite number of tokens before the `QUESTION_MARK` to determine
if comparison operators or similar conditions are being used. This is unsuitable for an array-based Lexer structure.

```console
$ ruby -rprism -e 'p Prism.lex("foo.bar==baz.qux.quux ? y : z").value.map(&:first).map(&:type)'
[:IDENTIFIER, :DOT, :IDENTIFIER, :EQUAL_EQUAL, :IDENTIFIER, :DOT, :IDENTIFIER, :DOT, :IDENTIFIER, :QUESTION_MARK, :IDENTIFIER, :COLON, :IDENTIFIER, :EOF]
```

Therefore, redundant cases may occur, but spaces will be uniformly added around the `?`.

On the other hand, if the RHG is a predicate method as shown below, the space before the `?` can be removed, and it will be done accordingly:

```ruby
foo==bar?? x: y # OK
```